+++ packages/nucleus/lib/src/atoms.dart
@@ -1,6 +1,6 @@
-export 'atoms/base.dart';
 export 'atoms/atom.dart';
-export 'atoms/derived.dart';
 export 'atoms/future.dart';
+export 'atoms/primitive.dart';
+export 'atoms/read_only.dart';
 export 'atoms/stream.dart';
 export 'atoms/managed.dart';