+++ packages/nucleus/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.3
+
+- Rename `DerivedAtom` to `ReadOnlyAtom`
+- Rename `Atom` to `PrimitiveAtom`
+- Rename `AtomBase` to `Atom`
+
 ## 0.0.2
 
 - Add initial project